+++ SlaveBoard/SlaveBoard.ino
@@ -1,5 +1,4 @@
 #include "I2C.h"
-#include "LiquidCrystal_I2C.h"
 
 #define ON true
 #define OFF false
@@ -11,11 +10,16 @@ const uint8_t m_size = 5;
 
 //corresponds to current state of LEDs in matrix, 1 = on
 bool state[m_size*m_size];
+
+//bool state[m_size*m_size] = { 0,1,1,0,1,
+//                              1,1,0,1,1,
+//                              0,0,1,0,0,
+//                              1,0,0,1,1,
+//                              1,1,1,1,1 };
                                   
 uint8_t cur_row = 0; 
 
 I2C_Slave slave;
-LiquidCrystal_I2C lcd(0x3f, 2, 1, 0, 4, 5, 6, 7, 3, POSITIVE);
 
 void eventhandle();
 
@@ -27,14 +31,12 @@ void setup() {
     digitalWrite(colpin[i], LOW);
   }
   for(int i = 0; i < m_size*m_size; ++i) state[i] = OFF;
-  lcd.begin(16,2);
   slave.init(eventhandle);  
 }
 
 void loop() {
-  //turn off the other two rows
-  digitalWrite(rowpin[(cur_row+1)%m_size], LOW); 
-  digitalWrite(rowpin[(cur_row+2)%m_size], LOW);
+  //turn off the other rows
+  for(int i = 1; i < m_size; ++i) digitalWrite(rowpin[(cur_row+i)%m_size], LOW);
 
   //column = LOW lets current pass, HIGH blocks current
   for(int col = 0; col < m_size; ++col) digitalWrite(colpin[col], !(state[(cur_row)*m_size+col]));
@@ -62,20 +64,17 @@ void eventhandle(){
     state[event-SET0] = ON;
   }
   else{
-    lcd.clear();
     switch(event){
       case WIN: 
-        lcd.write("You win");
         break;
       case RESET:
-        lcd.write("Resetting");
         for(byte i = 0; i < 25; ++i) state[i] = OFF;
         break;
       case M_ERROR: 
-        lcd.write("Error");
         for(byte i = 0; i < 25; ++i) state[i] = OFF;
         break;
-      default: lcd.write("Unexpected Error");
+      default: 
+        break;
     }  
   }
 }
